+++ openhasp_config_manager/openhasp_client/model/configuration/debug_config.py
