+++ openhasp_config_manager/openhasp_client/model/configuration/device_config.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from openhasp_config_manager.openhasp_client.model.screen_config import ScreenConfig
+from openhasp_config_manager.openhasp_client.model.configuration.screen_config import ScreenConfig
 
 
 @dataclass
