+++ openhasp_config_manager/openhasp_client/model/configuration/gui_config.py
