+++ openhasp_config_manager/openhasp_client/model/configuration/hasp_config.py
