+++ openhasp_config_manager/openhasp_client/model/configuration/http_config.py
