+++ openhasp_config_manager/openhasp_client/model/configuration/screen_config.py
