+++ openhasp_config_manager/openhasp_client/model/configuration/telnet_config.py
