+++ openhasp_config_manager/openhasp_client/model/configuration/wifi_config.py
