+++ openhasp_config_manager/openhasp_client/model/device.py
@@ -4,7 +4,7 @@ from typing import List
 
 from openhasp_config_manager.openhasp_client.model.component import JsonlComponent, CmdComponent, \
     ImageComponent, FontComponent
-from openhasp_config_manager.openhasp_client.model.config import Config
+from openhasp_config_manager.openhasp_client.model.configuration.config import Config
 
 
 @dataclass
