+++ openhasp_config_manager/openhasp_client/webservice_client.py
@@ -3,12 +3,12 @@ from typing import Dict, List
 import orjson
 import requests
 
-from openhasp_config_manager.openhasp_client.model.debug_config import DebugConfig
-from openhasp_config_manager.openhasp_client.model.gui_config import GuiConfig
-from openhasp_config_manager.openhasp_client.model.hasp_config import HaspConfig
-from openhasp_config_manager.openhasp_client.model.http_config import HttpConfig
-from openhasp_config_manager.openhasp_client.model.mqtt_config import MqttConfig, MqttTopicConfig
-from openhasp_config_manager.openhasp_client.model.telnet_config import TelnetConfig
+from openhasp_config_manager.openhasp_client.model.configuration.debug_config import DebugConfig
+from openhasp_config_manager.openhasp_client.model.configuration.gui_config import GuiConfig
+from openhasp_config_manager.openhasp_client.model.configuration.hasp_config import HaspConfig
+from openhasp_config_manager.openhasp_client.model.configuration.http_config import HttpConfig
+from openhasp_config_manager.openhasp_client.model.configuration.mqtt_config import MqttConfig, MqttTopicConfig
+from openhasp_config_manager.openhasp_client.model.configuration.telnet_config import TelnetConfig
 
 GET = "GET"
 POST = "POST"
