+++ openhasp_config_manager/processing/jsonl/__init__.py
@@ -1,7 +1,7 @@
 import abc
 from typing import Dict
 
-from openhasp_config_manager.openhasp_client.model.config import Config
+from openhasp_config_manager.openhasp_client.model.configuration.config import Config
 
 
 class JsonlObjectProcessor(metaclass=abc.ABCMeta):
